Move steps into a separate module
